requirements.txt
anyio==3.6.2
asgiref==3.6.0
click==8.1.3
Django==4.1.5
django-model-utils==4.3.1
graphql-core==3.2.3
h11==0.14.0
idna==3.4
libcst==0.4.9
markdown-it-py==2.1.0
mdurl==0.1.2
mypy-extensions==0.4.3
Pygments==2.14.0
python-dateutil==2.8.2
python-multipart==0.0.5
PyYAML==6.0
rich==13.2.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starlette==0.23.1
strawberry-graphql==0.154.1
strawberry-graphql-django==0.9
typing-inspect==0.8.0
typing_extensions==4.4.0
uvicorn==0.20.0
